refactor(frontend): split LoopStudioTriggersPanel and LoopStudioExecutionsPanel
Split LoopStudioTriggersPanel (825→1 line) into:
- triggers/helpers.tsx: TRIGGER_META constants
- triggers/CronConfigForm.tsx: cron configuration form
- triggers/TodoSelectorForm.tsx: todo selector for todo_completed/todo_state_changed triggers
- triggers/TagSelectorForm.tsx: tag selector for tag_added triggers
- triggers/FeishuMessageConfigForm.tsx: feishu message trigger config
- triggers/FeishuCommandConfigForm.tsx: feishu command trigger config
- triggers/TodoStateChangedConfigForm.tsx: todo state change config
- triggers/TriggerConfigContent.tsx: config content dispatcher
- triggers/index.tsx: main panel component

Split LoopStudioExecutionsPanel (918→1 line) into:
- executions/helpers.tsx: execStatusView, durationLabel, formatToken, formatCost
- executions/TokenSummaryBar.tsx: token summary display
- executions/BlackboardDrawer.tsx: blackboard drawer component
- executions/StepExecList.tsx: step execution list cards
- executions/index.tsx: main executions panel component

All exports maintained via re-export from original files for backward compatibility.
